voicevoxを使った処理をクラス化して使用するように変更
