+++ discord-app/src/discord_bot/voice_state_update/VoiceStateUpdate.ts
@@ -1,15 +1,7 @@
 const { Client, GuildMember, VoiceState } = require('discord.js');
-const { default: axios } = require('axios');
-
-const {
-  createAudioPlayer,
-  createAudioResource,
-  NoSubscriberBehavior,
-  StreamType,
-} = require('@discordjs/voice');
-
 const fs = require('fs');
 const DiscordBot = require('../DiscordBot').default;
+const VoiceVox = require('../../voice_vox/VoiceVox').default;
 
 /**
  * ボイスチャンネルに誰かが参加/退出した時に行う処理クラス
@@ -105,72 +97,8 @@ export default class VoiceStateUpdate {
 
     if (!fs.existsSync(filePath)) fs.mkdirSync(filePath, { recursive: true });
 
-    const readText = VoiceStateUpdate.formatMessage(announceVoice);
-    await VoiceStateUpdate.generateAudioFile(readText, wavFile, client.speakerId);
-    VoiceStateUpdate.play(wavFile, client.connection);
-  }
-
-  /**
-   * 読み上げるテキストの内容をフォーマット（絵文字やURLなどを排除）
-   *
-   * @param {string} messageContent フォーマット対象のメッセージ
-   *
-   * @return {string}
-   */
-  private static formatMessage(messageContent: string): string {
-    let formattedMessageContent: string = messageContent;
-
-    formattedMessageContent = formattedMessageContent.replace(/<:[a-zA-Z0-9_]+:[0-9]+>/g, '');
-    formattedMessageContent = formattedMessageContent.replace(/\r?\n/g, '、');
-
-    return formattedMessageContent;
-  }
-
-  /**
-   * 入力されたテキストを読み上げるwavファイルを生成
-   *
-   * @param {string} readText 読み上げ対象のメッセージ
-   * @param {string} wavFile wavファイルの保存先パス
-   * @param {string} speakerId
-   *
-   * @return {void}
-   */
-  private static async generateAudioFile(readText: string, wavFile: string, speakerId: string) {
-    const voiceVox = axios.create({ baseURL: 'http://voicevox-engine:50021/', proxy: false });
-
-    const audioQuery = await voiceVox.post(`audio_query?text=${encodeURI(readText)}&speaker=${speakerId}`, {
-      headers: { accept: 'application/json' },
-    });
-
-    const synthesis = await voiceVox.post(`synthesis?speaker=${speakerId}`, JSON.stringify(audioQuery.data), {
-      responseType: 'arraybuffer',
-      headers: {
-        accept: 'audio/wav',
-        'Content-Type': 'application/json',
-      },
-    });
-
-    fs.writeFileSync(wavFile, Buffer.from(synthesis.data), 'binary');
-  }
-
-  /**
-   * 生成したwavファイルを元に読み上げ
-   *
-   * @param {string} wavFile 生成したwavファイル
-   * @param {any} connection ボイスチャンネルオブジェクト
-   *
-   * @return {void}
-   */
-  private static play(wavFile: string, connection: any): void {
-    const resource = createAudioResource(wavFile, { inputType: StreamType.Arbitrary });
-    const player = createAudioPlayer({
-      behaviors: {
-        noSubscriber: NoSubscriberBehavior.Pause,
-      },
-    });
-
-    player.play(resource);
-
-    connection.subscribe(player);
+    const readText = VoiceVox.formatMessage(announceVoice);
+    await VoiceVox.generateAudioFile(readText, wavFile, client.speakerId);
+    VoiceVox.play(wavFile, client.connection);
   }
 }
